% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.addFormat}
\alias{ajv.addFormat}
\title{The Ajv.addFormat method}
\usage{
ajv.addFormat(this, key, format)
}
\arguments{
\item{key}{String; the name with format to add.}

\item{format}{the format to be added.  Note that JavaScript object literals
should be enclosed in a call to \code{\link[V8]{JS}}. (i.e.
\code{ajv$addFormat("numbers",JS("/^\\\\d+$/"))})}
}
\value{
\code{invisible(NULL)}
}
\description{
Add a string format to an Ajv instance.
}

