% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.addSchema}
\alias{ajv.addSchema}
\title{The Ajv.addSchema method}
\usage{
ajv.addSchema(this, schema, key)
}
\arguments{
\item{schema}{The schema to be added.
\code{schema} may be a  valid JSON string, an R object (i.e. `list(...)`), a
connection to a JSON file, or the name of a JSON or YAML file.  YAML files
are parsed via [js-yaml](https://www.npmjs.com/package/js-yaml)'s
`safeLoad()` method.}

\item{key}{String; the name with which to store the schema}
}
\value{
\code{invisible(NULL)}
}
\description{
The add a schema to an Ajv instance
}

