% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.keyword}
\alias{ajv.keyword}
\title{The Ajv.addFormat method}
\usage{
ajv.keyword(this, key, object)
}
\arguments{
\item{key}{String; the name with keyword to add.}

\item{object}{the format to be added.  Must be enclosed in a call to
\code{\link[V8]{JS}}.}
}
\value{
\code{invisible(NULL)}
}
\description{
Add a string format to an Ajv instance.
}

