% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.removeSchema}
\alias{ajv.removeSchema}
\title{The Ajv.removeSchema method}
\usage{
ajv.removeSchema(this, key)
}
\arguments{
\item{key}{String; the name with schema to remove}
}
\value{
\code{invisible(NULL)}
}
\description{
The remove a schema from an Ajv instance
}

